requirements.txt
asgiref==3.8.1
boto3==1.34.146
botocore==1.34.146
Django==5.0.7
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-indexer==0.3.0
django-js-asset==2.2.0
django-maintenance-mode==0.21.1
django-paging==0.2.5
django-prometheus==2.3.1
django-redis==5.4.0
django-sentry==1.13.5
django-storages==1.14.4
django-taggit==5.0.1
django-templatetag-sugar==1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
dnspython==2.6.1
docutils==0.21.2
eventlet==0.36.1
greenlet==3.0.3
jmespath==1.0.1
lockfile==0.12.2
logging==0.4.9.6
pillow==10.4.0
prometheus_client==0.20.0
psycopg2-binary==2.9.9
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fsutil==0.14.1
raven==6.10.0
redis==5.0.7
s3transfer==0.10.2
setuptools==71.1.0
six==1.16.0
South==1.0.2
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
